calibration: fix correspondence search
A big win here.

Previously I was constructing a list of 3D candidate points
close to the p,t location where the projector pixel was seen.
But I was iterating this list BACKWARDS (due to popping it).

The spiral iterator revealed this, and now we start much closer
to the initial guess, which means we succeed faster, and we hit
closer to the approximate search grid.